added local_minima_variation.py file
