+++ src/serena/utilities/ensemble_variation.py
@@ -18,7 +18,6 @@ from enum import Enum
 
 from serena.utilities.ensemble_structures import Sara2SecondaryStructure, Sara2StructureList
 from serena.utilities.ensemble_groups import MultipleEnsembleGroups
-#from serena.utilities.thread_manager import EV_Shuttle, EV_ThreadProcessor
 
 class SourceMFE(Enum):
     NONE = 0
@@ -214,109 +213,3 @@ class EnsembleVariation():
         token.group_dict[group_num] = result
         token.group_done_status[group_num] = True
 
-class EV_ThreadProcessor():
-    
-    def __init__(self, stuctures: List[Sara2StructureList], comp_structure: Sara2SecondaryStructure = Sara2SecondaryStructure(), comp_struct_list_option: List[Sara2SecondaryStructure] = []) -> None:
-        self._sara2_groups: List[Sara2StructureList] = stuctures
-        num_groups:int = len(stuctures)
-        self._num_groups: int =  num_groups
-        self._group_token: EV_Token = EV_Token(num_groups)
-        self._EV: EnsembleVariation = EnsembleVariation()
-        self._comparison_structure: Sara2SecondaryStructure = comp_structure
-
-    @property
-    def sara2_groups(self):
-        return self._sara2_groups
-
-    @sara2_groups.setter
-    def sara2_groups(self, new_list:List[Sara2StructureList]):
-        self._sara2_groups = new_list
-    
-    @property
-    def comparison_structure(self):
-        return self._comparison_structure
-
-    @comparison_structure.setter
-    def comparison_structure(self, new_struct:Sara2SecondaryStructure):
-        self._comparison_structure = new_struct
-    
-    @property
-    def comp_struct_list_option(self):
-        return self._comp_struct_list_option
-
-    @comp_struct_list_option.setter
-    def comp_struct_list_option(self, new_list:List[Sara2SecondaryStructure]):
-        self._comp_struct_list_option = new_list
-
-    @property
-    def num_groups(self):
-        return self._num_groups
-
-    @num_groups.setter
-    def num_groups(self, new_num:int):
-        self._num_groups = new_num
-
-    @property
-    def group_token(self):
-        return self._group_token
-
-    @group_token.setter
-    def group_token(self, new_token:EV_Token):
-        self._group_token = new_token
-    
-    @property
-    def EV(self):
-        return self._EV
-
-    @EV.setter
-    def EV(self, new_ev:EnsembleVariation):
-        self._EV = new_ev
-
-    def run_EV(self):
-        self.start_calculations()
-        self.wait_for_finish()
-        #the test should be done now
-        #check foor index that is -1 and if so then use prev value
-        num_groups:int = len(self.group_token.group_results)
-        for index in range(1, num_groups):
-            if self.group_token.group_results[index].ev_normalized == -1:
-                previous_EV = self.group_token.group_results[index-1]
-                self.group_token.group_results[index] = previous_EV
-                self.group_token.group_dict[index] = previous_EV
-        return self.group_token
-
-    def start_calculations(self):
-        comp_structure: Sara2SecondaryStructure = Sara2SecondaryStructure()               
-        for thread_index in range(self.num_groups):
-            if len(self.comp_struct_list_option) == self.num_groups:
-                comp_structure = self.comp_struct_list_option[thread_index]
-            else:
-                comp_structure = self.comparison_structure
-            sara2_structs: Sara2StructureList  = self.sara2_groups[thread_index]
-            new_shuttle: EV_Shuttle = EV_Shuttle(structs_list=sara2_structs, mfe=comp_structure, group_index=thread_index,token=self.group_token) 
-            mew_thread = threading.Thread(target=self.EV.thread_EV, args=[new_shuttle])
-            mew_thread.start()
-
-    
-    def wait_for_finish(self):
-                
-        stop:bool = False
-        while stop == False:
-            print(f'Checking LMV status at {datetime.now()}')
-            current_status: List[bool] = self.group_token.group_done_status
-            is_done:bool = self.group_token.is_done
-            
-            message: str = ''
-            for index in range(self.num_groups):
-                goup_value:str = self.group_token.group_values[index]
-                done_status: bool = self.group_token.group_done_status[index]
-                message = message + f'Group_{index+1}: kcal_group={goup_value}, status={done_status}\n'
-            print(message)
-
-            if is_done == True:
-                stop = True
-                print(f'Its done at {datetime.now()}')
-            else:
-                dwell_time:int = 5
-                print(f'dwelling for {dwell_time} seconds until next check')
-                time.sleep(dwell_time)
